# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Description of the API",
        "title": "WB Author API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/v1/get/mocks": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a list of Subscriptions based on provided filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Get a list of Okx",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Okx"
                            }
                        }
                    }
                }
            }
        },
        "/v1/get/okx-info": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return Okx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Get Okx",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Okx"
                        }
                    }
                }
            }
        },
...
        "/v1/update/okx/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Okx",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Okx"
                ],
                "summary": "Update Okx",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Okx ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Okx",
                        "name": "Okx",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/j-ai-trade_modules_okx_model_dto.SubscriptionUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Okx"
                        }
                    }
                }
            }
        },
...
        "/v2/Permission/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new Permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Create new Permission",
                "parameters": [
                    {
                        "description": "Create Permission",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Permission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Permission"
                        }
                    }
                }
            }
        },
        "/v2/Permission/delete": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a new Permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Delete new Permission",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Permission"
                        }
                    }
                }
            }
        },
        "/v2/Permission/get": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Return Permission",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Get Permission",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "User UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Permission"
                        }
                    }
                }
            }
        },
        "/v2/Permission/list": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve a list of Permissions based on provided filters and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Get a list of Permission",
                "parameters": [
                    {
                        "description": "Get Permissions",
                        "name": "Permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PermissionGetListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Permission"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.PaginationRequest": {
            "type": "object",
            "properties": {
                "index": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
        "j-ai-trade_modules_okx_model_dto.SubscriptionUpdateRequest": {
            "type": "object"
        },
...
        "model.Okx": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "model.Permission": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                }
            }
        },
        "model.PermissionGetListRequest": {
            "type": "object",
            "properties": {
                "Pagination": {
                    "$ref": "#/definitions/common.PaginationRequest"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}