# modules/okx/transport/gin/get_okx_info_handler.go
package ginNovel

import (
	"j-ai-trade/common"
	"j-ai-trade/modules/okx/biz"
	dto "j-ai-trade/modules/okx/model/dto"
	"j-ai-trade/modules/okx/storage"
	"net/http"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

// GetSubscription godoc
// @Summary Get Okx
// @Description Return Okx
// @Produce application/json
// @Tags Okx
// @Success 200 {object} model.Okx
// @securityDefinitions.apiKey token
// @in header
// @name Authorization
// @Security Bearer
// @Router /v1/get/okx-info [get]
func GetOkxInfo(db *gorm.DB) func(*gin.Context) {
	return func(c *gin.Context) {
		store := storage.NewPostgresStore(db)
		business := biz.NewGetOkxInfoBiz(store)

		data, err := business.GetOkxInfo(c.Request.Context())
		if err != nil {
			c.JSON(http.StatusBadRequest, common.BaseApiResponse[any]{
				HttpRequestStatus: http.StatusBadRequest,
				Success:           false,
				Message:           err.Error(),
				Data:              nil,
			})
			return
		}

		c.JSON(http.StatusOK, common.BaseApiResponse[dto.OkxInfoResponse]{
			Success:           true,
			HttpRequestStatus: http.StatusOK,
			Message:           "Okx info successfully",
			Data:              *data,
		})
	}
}