# modules/okx/transport/gin/update_okx_handler.go
package ginNovel

import (
	"j-ai-trade/common"
	"j-ai-trade/logger"
	"j-ai-trade/middlewares"
	"j-ai-trade/modules/okx/biz"
	model "j-ai-trade/modules/okx/model"
	dto "j-ai-trade/modules/okx/model/dto"
	"j-ai-trade/modules/okx/storage"
	"net/http"

	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

// UpdateSubscription godoc
// @Summary Update Okx
// @Description Update Okx
// @Param id path string true "Okx ID"
// @Param Okx body dto.SubscriptionUpdateRequest true "Update Okx"
// @Produce application/json
// @Tags Okx
// @Success 200 {object} model.Okx
// @securityDefinitions.apiKey token
// @in header
// @name Authorization
// @Security Bearer
// @Router /v1/update/okx/{id} [put]
func UpdateSubscription(db *gorm.DB) func(*gin.Context) {
	return func(c *gin.Context) {
		log := logger.GetLogger("UpdateSubscription", c.GetString(middlewares.RequestIDKey))
		id := c.Param("id")
		if id == "" {
			c.JSON(http.StatusBadRequest, common.BaseApiResponse[any]{
				HttpRequestStatus: http.StatusBadRequest,
				Success:           false,
				Message:           "id is required",
				Data:              nil,
			})
			log.Error().Msg("id is required")
			return
		}

		var input dto.SubscriptionUpdateRequest
		if err := c.ShouldBindJSON(&input); err != nil {
			c.JSON(http.StatusBadRequest, common.BaseApiResponse[any]{
				HttpRequestStatus: http.StatusBadRequest,
				Success:           false,
				Message:           err.Error(),
				Data:              nil,
			})
			log.Error().Msg("failed to bind request")
			return
		}

		data := model.Okx{}

		store := storage.NewPostgresStore(db)
		business := biz.NewUpdateSubscriptionBiz(store)

		if err := business.UpdateSubscription(c.Request.Context(), id, &data); err != nil {
			c.JSON(http.StatusBadRequest, common.BaseApiResponse[any]{
				HttpRequestStatus: http.StatusBadRequest,
				Success:           false,
				Message:           err.Error(),
				Data:              nil,
			})
			log.Error().Err(err).Str("request_id", c.GetString(middlewares.RequestIDKey)).Msg("failed to update Okx")
			return
		}

		c.JSON(http.StatusOK, common.BaseApiResponse[model.Okx]{
			Success:           true,
			HttpRequestStatus: http.StatusOK,
			Message:           "Okx is updated successfully",
			Data:              data,
		})
		log.Info().Str("Subscription_id", id).Msg("Okx is updated successfully")
	}
}